+++ HISTORY.md
@@ -1,5 +1,5 @@
 ## 0.8.1 (in progress)
-- Support joint recalling with GATK HapolotypeCaller and FreeBayes. The
+- Support joint recalling with GATK HapolotypeCaller, FreeBayes and Platypus. The
   `jointcaller` configuration variable enables calling concurrently in large
   populations by independently calling on samples them combining into a final
   combined callset with no-call/reference calls at any position called
@@ -7,6 +7,8 @@
 - Add qsignature tool to standard analysis
 - Support the SNAP aligner, which provides improved speed on systems with
   larger amount of memory (64Gb for human genome alignment).
+- Support the Platypus haplotype based variant caller for germline samples with
+  both batched and joint calling.
 - Enable structural variant analyses on shared batches (two tumors with same
   normal). Thanks to Miika Ahdesmaki.
 - Fix GATK version detection when `_JAVA_OPTIONS` specified. Thanks to Miika